deepfake_backend/requirements.txt
Flask
flask-cors
flask-pymongo
pymongo
werkzeug
fpdf2
gunicorn
python-dotenv
